% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{gtxlsx-package}
\alias{gtxlsx}
\alias{gtxlsx-package}
\title{gtxlsx: Write 'gt' and HTML Tables into 'openxlsx2' Workbooks}
\description{
Turns a 'gt' table into a range of cells in an 'openxlsx2' workbook, keeping the heading, column spanners, row groups, stub, summary rows, footnotes and the styling set through 'gt'. Numbers stay numbers wherever a spreadsheet number format can reproduce what 'gt' shows. A second entry point does the same for a plain HTML table, so output from other table packages can be written to a worksheet as well; that path needs nothing beyond 'openxlsx2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://janmarvin.github.io/gtxlsx/}
  \item \url{https://github.com/JanMarvin/gtxlsx}
  \item Report bugs at \url{https://github.com/JanMarvin/gtxlsx/issues}
}

}
\author{
\strong{Maintainer}: Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Authors:
\itemize{
  \item Jan Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}
}

}
\keyword{internal}
